+++ test/completion/init.lua
@@ -2060,6 +2060,23 @@ TEST [[
 ---@param x cc
 local function f(x) end
 
+f({aaa = 1,$})
+]]
+{
+    {
+        label = 'bbb',
+        kind  = define.CompletionItemKind.Property,
+    },
+}
+
+TEST [[
+---@class cc
+---@field aaa number # a1
+---@field bbb number # a2
+
+---@param x cc
+local function f(x) end
+
 f({
     {
         $